Add validations for company and addresses in the request

+++ db/migrate/20241223121257_create_companies.rb
@@ -2,7 +2,7 @@ class CreateCompanies < ActiveRecord::Migration[8.0]
   def change
     create_table :companies do |t|
       t.string :name
-      t.integer :registration_number
+      t.bigint :registration_number
 
       t.timestamps
     end
